# rent-a-ride-ui/src/app/services/models/borrowed-vehicle-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BorrowedVehicleResponse {
  carModel?: string;
  id?: number;
  licensePlate?: string;
  ownerName?: string;
  rate?: number;
  returnApproved?: boolean;
  returned?: boolean;
}