# rent-a-ride-ui/src/app/services/models/page-response-borrowed-vehicle-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BorrowedVehicleResponse } from '../models/borrowed-vehicle-response';
export interface PageResponseBorrowedVehicleResponse {
  content?: Array<BorrowedVehicleResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}