# rent-a-ride-ui/src/app/services/models/page-response-vehicle-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { VehicleResponse } from '../models/vehicle-response';
export interface PageResponseVehicleResponse {
  content?: Array<VehicleResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}