# rent-a-ride-ui/src/app/services/models/registeration-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface RegisterationRequest {
  email: string;
  firstname?: string;
  lastname?: string;
  password: string;
}