# rent-a-ride-ui/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification - Balamurugan","description":"OpenApi documentaion for Springboot application of a vehicle renting website ","contact":{"name":"Balamurugan","email":"rbalamurgan1000@gmail.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8088/api/v1","description":"Local Environment"},{"url":"https://vehicleRent/api/v1","description":"Production Environment"}],"security":[{"bearerAuthentication":[]}],"paths":{"/vehicles":{"get":{"tags":["Vehicle"],"operationId":"findAllVehicles","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseVehicleResponse"}}}}}},"post":{"tags":["Vehicle"],"operationId":"saveVehicle","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/vehicles/borrow/{vehicle-id}":{"post":{"tags":["Vehicle"],"operationId":"borrowVehicle","parameters":[{"name":"vehicle-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/feedbacks":{"post":{"tags":["Feedback"],"operationId":"saveFeedback","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/vehicles/shareable/{vehicle-id}":{"patch":{"tags":["Vehicle"],"operationId":"updateShaeableStatus","parameters":[{"name":"vehicle-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/vehicles/borrow/return/{vehicle-id}":{"patch":{"tags":["Vehicle"],"operationId":"returnVehicle","parameters":[{"name":"vehicle-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/vehicles/borrow/return/approve/{vehicle-id}":{"patch":{"tags":["Vehicle"],"operationId":"approveReturnVehicle","parameters":[{"name":"vehicle-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/vehicles/archived/{vehicle-id}":{"patch":{"tags":["Vehicle"],"operationId":"updateArchivedStatus","parameters":[{"name":"vehicle-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/vehicles/{vehicle-id}":{"get":{"tags":["Vehicle"],"operationId":"findVehicleById","parameters":[{"name":"vehicle-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleResponse"}}}}}}},"/vehicles/returned":{"get":{"tags":["Vehicle"],"operationId":"findAllReturnedVehicles","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseBorrowedVehicleResponse"}}}}}}},"/vehicles/owner":{"get":{"tags":["Vehicle"],"operationId":"findAllVehiclesByOwner","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseVehicleResponse"}}}}}}},"/vehicles/borrowed":{"get":{"tags":["Vehicle"],"operationId":"findAllBorrowedVehicles","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseBorrowedVehicleResponse"}}}}}}},"/feedbacks/vehicle/{vehicle-id}":{"get":{"tags":["Feedback"],"operationId":"getAllFeedbackByVehicleId","parameters":[{"name":"vehicle-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseFeedbackResponse"}}}}}}},"/auth/activate-account":{"get":{"tags":["Authentication"],"operationId":"confirm","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"VehicleRequest":{"required":["carModel","description","licensePlate","ownerName","phone"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"carModel":{"type":"string"},"ownerName":{"type":"string"},"licensePlate":{"type":"string"},"description":{"type":"string"},"phone":{"type":"string"},"price":{"type":"string"},"cover":{"type":"string"},"shareable":{"type":"boolean"}}},"FeedbackRequest":{"required":["comment","vehicleId"],"type":"object","properties":{"note":{"maximum":5,"minimum":0,"type":"number","format":"double"},"comment":{"type":"string"},"vehicleId":{"type":"integer","format":"int32"}}},"RegisterationRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"}}},"AuthenticationRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}},"PageResponseVehicleResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/VehicleResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"VehicleResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"carModel":{"type":"string"},"ownerName":{"type":"string"},"licensePlate":{"type":"string"},"description":{"type":"string"},"phone":{"type":"string"},"price":{"type":"string"},"cover":{"type":"string"},"rate":{"type":"number","format":"double"},"archived":{"type":"boolean"},"shareable":{"type":"boolean"}}},"BorrowedVehicleResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"carModel":{"type":"string"},"ownerName":{"type":"string"},"licensePlate":{"type":"string"},"rate":{"type":"number","format":"double"},"returned":{"type":"boolean"},"returnApproved":{"type":"boolean"}}},"PageResponseBorrowedVehicleResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/BorrowedVehicleResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"FeedbackResponse":{"type":"object","properties":{"note":{"type":"number","format":"double"},"comment":{"type":"string"},"ownFeedback":{"type":"boolean"}}},"PageResponseFeedbackResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/FeedbackResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}}},"securitySchemes":{"bearerAuthentication":{"type":"http","description":"JWT authentication","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}